# api/swagger.yaml
openapi: 3.0.3
info:
  title: GroupScout API
  description: API for managing lead collection, enrichment, and notification.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Lead Generation Server (default)
  - url: http://localhost:8081
    description: Alerting Service (alertd)
paths:
...
  /run:
    post:
      summary: Trigger pipeline
      description: Manually triggers the collect-enrich-notify pipeline.
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                bcbid_raw_input:
                  type: string
                  description: Optional raw input for BC Bid collector.
      responses:
        '200':
          description: Pipeline completed successfully
        '401':
          description: Unauthorized
        '500':
          description: Pipeline execution failed
  /digest:
    post:
      summary: Trigger weekly digest
      description: Sends a weekly digest of leads via email.
      security:
        - BearerAuth: []
      parameters:
        - name: to
          in: query
          description: Recipient email address (defaults to alvin@groupscout.ai)
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Digest sent successfully
        '401':
          description: Unauthorized
        '500':
          description: Failed to send digest
  /n8n/webhook:
    post:
      summary: Receive lead from n8n
      description: Endpoint for n8n to push new leads into the system.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lead'
      responses:
        '201':
          description: Lead created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  id:
                    type: string
        '400':
          description: Invalid JSON body
        '401':
          description: Unauthorized
        '500':
          description: Failed to store lead
...
  /api/leads:
    get:
      summary: List UI leads
      description: Returns filtered lead summaries for the operator inbox.
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: source
          in: query
          schema:
            type: string
        - name: min_score
          in: query
          schema:
            type: integer
            minimum: 0
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lead summaries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeadListResponse'
        '400':
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        '500':
          description: Failed to list leads
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: groupscout_session
...
  schemas:
    APIError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LeadSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        source:
          type: string
        location:
          type: string
        project_value:
          type: integer
          format: int64
        priority_score:
          type: integer
        priority_reason:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        has_raw:
          type: boolean
        audit_source_url:
          type: string
          format: uri
      required:
        - id
        - title
        - source
        - location
        - project_value
        - priority_score
        - priority_reason
        - status
        - created_at
        - updated_at
        - has_raw
    LeadListFilters:
      type: object
      properties:
        status:
          type: string
        source:
          type: string
        min_score:
          type: integer
        q:
          type: string
        limit:
          type: integer
        cursor:
          type: string
    LeadListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/LeadSummary'
        next_cursor:
          type: string
        filters:
          $ref: '#/components/schemas/LeadListFilters'
      required:
        - items
        - next_cursor
        - filters
...
    Lead:
      type: object
      properties:
        ID:
          type: string
          format: uuid
        RawProjectID:
          type: string
        Source:
          type: string
        Title:
          type: string
        Location:
          type: string
        ProjectValue:
          type: integer
          format: int64
        GeneralContractor:
          type: string
        Applicant:
          type: string
        Contractor:
          type: string
        SourceURL:
          type: string
          format: uri
        ProjectType:
          type: string
        EstimatedCrewSize:
          type: integer
        EstimatedDurationMonths:
          type: integer
        OutOfTownCrewLikely:
          type: boolean
        PriorityScore:
          type: integer
        PriorityReason:
          type: string
        SuggestedOutreachTiming:
          type: string
        Notes:
          type: string
        Status:
          type: string
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time